Improve implicit typing of state
This swaps the order of overloads so the primary signature is `setState(Partial<TState>)`. This means that haxe can better infer TState's field types when calling set state. For example, this now works
```haxe
setState({
    someEnum: SomeEnumValue
});
```

Whereas before it would have failed with "Unknown identifier : SomeEnumValue" and you'd need to tell haxe that this field has type SomeEnum. So the workaround was
```haxe
setState({
    someEnum: SomeEnum.SomeEnumValue
});
```